Add `FlexCarouselInner` component
