+++ app/views/kitten/components/carousel/flex-carousel.html.erb
@@ -1,15 +1,15 @@
 <div class="k-Grid">
   <div class="k-Grid__col k-Grid__col--12">
     <%= react_component 'KarlFlexCarousel', props: {
-      content: [
-        {
-          title: "Profitez de nos conseils en investissement",
-          titleLineBreak: "grâce au Mode Complet",
-          paragraph: "Nous vous conseillons de façon personnalisée en fonction de  votre situation",
-          paragraphLineBreak: "et de vos préférences d'investissements",
-          button: "En savoir plus",
-        }
-      ]
+      # content: [
+      #   {
+      #     title: "Profitez de nos conseils en investissement",
+      #     titleLineBreak: "grâce au Mode Complet",
+      #     paragraph: "Nous vous conseillons de façon personnalisée en fonction de  votre situation",
+      #     paragraphLineBreak: "et de vos préférences d'investissements",
+      #     button: "En savoir plus",
+      #   }
+      # ]
     } %>
   </div>
 </div>
